Load Screens
Created a load screen animation widget to be run whenever the kuching or belawai levels are being launched.  (an earth with an orbiting moon), just to clear up some of the initial loading of the map and the initial brightness flares from cesiums sun

refactored some of the blueprints, begun labelling subsections of logic in blueprints for more clarity moving forward.